backend/pyproject.toml
[project]
name = "secretly-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.17.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "python-nanoid>=2.0.0",
]

[dependency-groups]
dev = [
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.24.0",
]
